+++ src/mbplsr.jl
@@ -1,6 +1,8 @@
 struct Mbplsr
     fm
     T::Matrix{Float64}
+    R::Matrix{Float64}
+    C::Matrix{Float64}
     bscales::Vector{Float64}
     xmeans::Vector{Vector{Float64}}
     xscales::Vector{Vector{Float64}}
@@ -11,10 +13,10 @@ end
 
 """
     mbplsr(Xbl, Y, weights = ones(nro(Xbl[1])); nlv, 
-        bscal = "frob", scal = false)
+        bscal = "none", scal = false)
     mbplsr!(Xbl, Y, weights = ones(nro(Xbl[1])); nlv, 
-        bscal = "frob", scal = false)
-Multiblock PLSR (MBPLSR).
+        bscal = "none", scal = false)
+Multiblock PLSR (MBPLSR) - Fast version.
 * `Xbl` : List (vector) of blocks (matrices) of X-data. 
     Each component of the list is a block.
 * `Y` : Y-data.
@@ -28,7 +30,8 @@ Multiblock PLSR (MBPLSR).
     of `Y` is scaled by its uncorrected standard deviation 
     (before the block scaling).
 
-PLSR (X, Y) where X is the horizontal concatenation of the blocks in `Xbl`.
+This is the PLSR (X, `Y`) where X is the horizontal concatenation of the blocks in `Xbl`.
+The function gives the same results as function `mbplswest`.
 
 ## Examples
 ```julia
@@ -46,7 +49,7 @@ Xbl = mblock(X, listbl)
 # "New" = first two rows of Xbl 
 Xbl_new = mblock(X[1:2, :], listbl)
 
-bscal = "frob"
+bscal = "none"
 nlv = 5
 fm = mbplsr(Xbl, y; nlv = nlv, bscal = bscal) ;
 pnames(fm)
@@ -54,11 +57,13 @@ fm.T
 Jchemo.transform(fm, Xbl_new)
 [y Jchemo.predict(fm, Xbl).pred]
 Jchemo.predict(fm, Xbl_new).pred
+
+summary(fm, Xbl)
 ```
 """
 
 function mbplsr(Xbl, Y, weights = ones(nro(Xbl[1])); nlv, 
-        bscal = "frob", scal = false)
+        bscal = "none", scal = false)
     nbl = length(Xbl)  
     zXbl = list(nbl, Matrix{Float64})
     @inbounds for k = 1:nbl
@@ -69,7 +74,7 @@ function mbplsr(Xbl, Y, weights = ones(nro(Xbl[1])); nlv,
 end
 
 function mbplsr!(Xbl, Y, weights = ones(nro(Xbl[1])); nlv, 
-        bscal = "frob", scal = false)
+        bscal = "none", scal = false)
     nbl = length(Xbl)
     Y = ensure_mat(Y)
     q = nco(Y)
@@ -106,53 +111,39 @@ function mbplsr!(Xbl, Y, weights = ones(nro(Xbl[1])); nlv,
     end
     X = reduce(hcat, Xbl)
     fm = plskern(X, Y, weights; nlv = nlv, scal = false)
-    Mbplsr(fm, fm.T, bscales, xmeans, xscales, ymeans, yscales, weights)
-end
-
-""" 
-    transform(object::Mbplsr, Xbl; nlv = nothing)
-Compute latent variables (LVs = scores T) from a fitted model.
-* `object` : The fitted model.
-* `Xbl` : A list (vector) of blocks (matrices) for which LVs are computed.
-* `nlv` : Nb. LVs to consider.
-""" 
-function transform(object::Mbplsr, Xbl; nlv = nothing)
-    nbl = length(Xbl)
-    zXbl = list(nbl, Matrix{Float64})
-    Threads.@threads for k = 1:nbl
-        zXbl[k] = cscale(Xbl[k], object.xmeans[k], object.xscales[k])
-    end
-    res = blockscal(zXbl, object.bscales)
-    X = reduce(hcat, res.X)
-    transform(object.fm, X; nlv = nlv)
+    Mbplsr(fm, fm.T, fm.R, fm.C, 
+        bscales, xmeans, xscales, ymeans, yscales, weights)
 end
 
 """
-    predict(object::Mbplsr, Xbl; nlv = nothing)
-Compute Y-predictions from a fitted model.
+    summary(object::Mbplsr, Xbl)
+Summarize the fitted model.
 * `object` : The fitted model.
-* `Xbl` : A list (vector) of X-data for which predictions are computed.
-* `nlv` : Nb. LVs, or collection of nb. LVs, to consider. 
+* `Xbl` : The X-data that was used to fit the model.
 """ 
-function predict(object::Mbplsr, Xbl; nlv = nothing)
+function Base.summary(object::Mbplsr, Xbl)
+    n, nlv = size(object.T)
     nbl = length(Xbl)
+    sqrtw = sqrt.(object.weights)
     zXbl = list(nbl, Matrix{Float64})
     Threads.@threads for k = 1:nbl
         zXbl[k] = cscale(Xbl[k], object.xmeans[k], object.xscales[k])
     end
-    res = blockscal(zXbl, object.bscales)
-    X = reduce(hcat, res.X)
-    W = Diagonal(object.yscales)
-    isnothing(nlv) ? le_nlv = 1 : le_nlv = length(nlv) 
-    pred = predict(object.fm, X; nlv = nlv).pred
-    if le_nlv == 1
-        pred = object.ymeans' .+ pred * W 
-    else
-        # Threads not faster
-        @inbounds for i = 1:le_nlv
-            pred[i] = object.ymeans' .+ pred[i] * W
-        end
+    zXbl = blockscal(zXbl, object.bscales).X
+    @inbounds for k = 1:nbl
+        zXbl[k] .= sqrtw .* zXbl[k]
     end
-    (pred = pred,)
+    # Explained_X
+    ssk = zeros(nbl)
+    @inbounds for k = 1:nbl
+        ssk[k] = ssq(zXbl[k])
+    end
+    sstot = sum(ssk)
+    tt = object.fm.TT
+    tt_adj = vec(sum(object.fm.P.^2, dims = 1)) .* tt
+    pvar = tt_adj / sstot
+    cumpvar = cumsum(pvar)
+    xvar = tt_adj / n    
+    explvarx = DataFrame(nlv = 1:nlv, var = xvar, pvar = pvar, cumpvar = cumpvar)     
+    (explvarx = explvarx,)
 end
-